refactor: renames subProviders to provider
